Move all -core unit tests under one executable
 * move source files
 * adapt cmake lists
 * few fixes in OlpClient tests
 * update CI scripts

Relates-to: OLPEDGE-591
