+++ olp-cpp-sdk-core/CMakeLists.txt
@@ -326,9 +326,5 @@ install (FILES ${GEOTYPE_HEADER} DESTINATION ${INCLUDE_DIRECTORY}/olp/core/geo)
 export_config()
 
 if(EDGE_SDK_ENABLE_TESTING)
-    add_subdirectory(tests/cache)
-    add_subdirectory(tests/client)
-    add_subdirectory(tests/olpclient)
-    add_subdirectory(tests/geo)
-    add_subdirectory(tests/thread)
+    add_subdirectory(tests)
 endif()
