+++ olp-cpp-sdk-core/tests/geo/coordinates/GeoPointTest.cpp
