+++ scripts/linux/fv/gitlab-olp-cpp-sdk-core-test.sh
@@ -1,0 +1,6 @@
+#!/bin/bash -e
+
+echo ">>> Core Test ... >>>"
+source $FV_HOME/olp-cpp-sdk-core-test.variables
+$REPO_HOME/build/olp-cpp-sdk-core/tests/olp-cpp-sdk-core-tests \
+    --gtest_output="xml:$REPO_HOME/reports/olp-core-test-report.xml"
