--- scripts/linux/fv/olp-cpp-sdk-core-cache-test.variables
@@ -1,1 +1,0 @@
-#to be added soon if required
