--- scripts/linux/fv/olp-cpp-sdk-core-client-test.variables
@@ -1,3 +1,0 @@
-export secret="${sdk_client_test_secret}" #defined previously as protected value on Gitlab
-export appid="${sdk_client_test_appid}" #defined previously as protected value on Gitlab
-export catalog="hrn:here:data:::hereos-internal-test-v2"
