--- scripts/linux/fv/olp-cpp-sdk-core-geo-test.variables
@@ -1,1 +1,0 @@
-#to be added soon if required
